requirements.txt
appnope==0.1.0
bleach==2.0.0
decorator==4.0.11
Django==1.11
django-filter==1.0.2
django-markdown==0.8.4
djangorestframework==3.6.2
greenlet==0.4.12
html5lib==0.999999999
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
msgpack-python==0.4.8
neovim==0.1.13
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
